# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: competition
    description: Competition operations
  - name: auth
    description: Authentication operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Trainer login with username and PIN
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Session token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResult"
        "401":
          description: Invalid credentials
  /competition/card/{token}:
    get:
      operationId: getPlayerCard
      tags: [competition]
      summary: Get a player's public virtual card by token (QR target)
      parameters:
        - name: token
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Player card
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Player"
        "404":
          description: Not found
  /competition/dashboard:
    get:
      operationId: getCompetitionDashboard
      tags: [competition]
      summary: Get competition dashboard
      responses:
        "200":
          description: Competition dashboard
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Dashboard"
  /competition/teams:
    get:
      operationId: listTeams
      tags: [competition]
      summary: List teams with roster summaries
      responses:
        "200":
          description: Teams
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Team"
  /competition/players:
    get:
      operationId: listPlayers
      tags: [competition]
      summary: List players and card statistics
      responses:
        "200":
          description: Players
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Player"
  /competition/teams/{id}:
    patch:
      operationId: updateTeam
      tags: [competition]
      summary: Edit a team (e.g. rename)
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTeamInput"
      responses:
        "200":
          description: Updated team
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Team"
  /competition/players/{id}:
    patch:
      operationId: updatePlayer
      tags: [competition]
      summary: Edit a player (rename or transfer to another team)
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdatePlayerInput"
      responses:
        "200":
          description: Updated player
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Player"
  /competition/settings:
    get:
      operationId: getSettings
      tags: [competition]
      summary: Get competition settings
      responses:
        "200":
          description: Settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Settings"
    patch:
      operationId: updateSettings
      tags: [competition]
      summary: Update competition settings (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateSettingsInput"
      responses:
        "200":
          description: Updated settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Settings"
  /competition/matches:
    get:
      operationId: listMatches
      tags: [competition]
      summary: List season matches
      responses:
        "200":
          description: Matches
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Match"
    post:
      operationId: createMatch
      tags: [competition]
      summary: Add a match (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateMatchInput"
      responses:
        "201":
          description: Created match
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Match"
  /competition/matches/{id}:
    patch:
      operationId: updateMatch
      tags: [competition]
      summary: Edit a match (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateMatchInput"
      responses:
        "200":
          description: Updated match
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Match"
    delete:
      operationId: deleteMatch
      tags: [competition]
      summary: Delete a match (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
  /competition/leaderboard:
    get:
      operationId: getLeaderboard
      tags: [competition]
      summary: Get player and team leaderboard
      responses:
        "200":
          description: Leaderboard
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Leaderboard"
  /competition/questions:
    get:
      operationId: listQuestions
      tags: [competition]
      summary: List challenge questions
      responses:
        "200":
          description: Questions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Question"
    post:
      operationId: createQuestion
      tags: [competition]
      summary: Add a challenge question
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateQuestionInput"
      responses:
        "201":
          description: Created question
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Question"
  /competition/events:
    post:
      operationId: recordCompetitionEvent
      tags: [competition]
      summary: Record a score-changing competition event
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CompetitionEventInput"
      responses:
        "201":
          description: Updated player card
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Player"
components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    Team:
      type: object
      required: [id, name, shortName, color, coach, assistantCoach, points, wins, draws, losses, goalsFor, goalsAgainst, playerCount, captainName]
      properties:
        id: { type: integer }
        name: { type: string }
        shortName: { type: string }
        color: { type: string }
        coach: { type: string }
        assistantCoach: { type: string }
        points: { type: integer }
        wins: { type: integer }
        draws: { type: integer }
        losses: { type: integer }
        goalsFor: { type: integer }
        goalsAgainst: { type: integer }
        playerCount: { type: integer }
        captainName: { type: string }
    PlayerStats:
      type: object
      required: [newPoints, repeatPoints, challengePoints, attendancePoints, onlinePoints, specialTaskPoints, total, redCards, manOfMatch]
      properties:
        newPoints: { type: integer }
        repeatPoints: { type: integer }
        challengePoints: { type: integer }
        attendancePoints: { type: integer }
        onlinePoints: { type: integer }
        specialTaskPoints: { type: integer }
        total: { type: integer }
        redCards: { type: integer }
        manOfMatch: { type: boolean }
    Player:
      type: object
      required: [id, name, teamId, teamName, role, avatarInitials, cardRating, cardToken, attributes, stats]
      properties:
        id: { type: integer }
        name: { type: string }
        teamId: { type: integer }
        teamName: { type: string }
        role: { type: string }
        avatarInitials: { type: string }
        cardRating: { type: integer }
        cardToken: { type: string }
        attributes:
          $ref: "#/components/schemas/PlayerAttributes"
        stats:
          $ref: "#/components/schemas/PlayerStats"
    PlayerAttributes:
      type: object
      required: [pace, defense, shooting, dribbling, passing, physical]
      properties:
        pace: { type: integer }
        defense: { type: integer }
        shooting: { type: integer }
        dribbling: { type: integer }
        passing: { type: integer }
        physical: { type: integer }
    Match:
      type: object
      required: [id, week, status, dateLabel, homeTeamId, homeTeamName, homeTeamShortName, awayTeamId, awayTeamName, awayTeamShortName, homeScore, awayScore, attacks]
      properties:
        id: { type: integer }
        week: { type: integer }
        status: { type: string }
        dateLabel: { type: string }
        homeTeamId: { type: integer }
        homeTeamName: { type: string }
        homeTeamShortName: { type: string }
        awayTeamId: { type: integer }
        awayTeamName: { type: string }
        awayTeamShortName: { type: string }
        homeScore: { type: integer }
        awayScore: { type: integer }
        attacks:
          type: array
          items:
            $ref: "#/components/schemas/Attack"
    Attack:
      type: object
      required: [id, label, description, status, winnerTeamId]
      properties:
        id: { type: integer }
        label: { type: string }
        description: { type: string }
        status: { type: string }
        winnerTeamId: { type: ["integer", "null"] }
    Dashboard:
      type: object
      required: [seasonLabel, completedWeeks, totalWeeks, totalStudents, activePlayers, nextSessionLabel, nextSessionDate, featuredPlayer, teams, recentActivity]
      properties:
        seasonLabel: { type: string }
        completedWeeks: { type: integer }
        totalWeeks: { type: integer }
        totalStudents: { type: integer }
        activePlayers: { type: integer }
        nextSessionLabel: { type: string }
        nextSessionDate: { type: string }
        featuredPlayer:
          $ref: "#/components/schemas/Player"
        teams:
          type: array
          items:
            $ref: "#/components/schemas/Team"
        recentActivity:
          type: array
          items:
            $ref: "#/components/schemas/Activity"
    Activity:
      type: object
      required: [id, title, description, timeLabel, type]
      properties:
        id: { type: integer }
        title: { type: string }
        description: { type: string }
        timeLabel: { type: string }
        type: { type: string }
    Leaderboard:
      type: object
      required: [teams, players]
      properties:
        teams:
          type: array
          items:
            $ref: "#/components/schemas/Team"
        players:
          type: array
          items:
            $ref: "#/components/schemas/Player"
    Question:
      type: object
      required: [id, category, prompt, points, difficulty, isPublished]
      properties:
        id: { type: integer }
        category: { type: string }
        prompt: { type: string }
        points: { type: integer }
        difficulty: { type: string }
        isPublished: { type: boolean }
    LoginInput:
      type: object
      required: [username, pin]
      properties:
        username: { type: string, minLength: 1 }
        pin: { type: string, minLength: 1 }
    LoginResult:
      type: object
      required: [token, name, username, role, teamId]
      properties:
        token: { type: string }
        name: { type: string }
        username: { type: string }
        role: { type: string, enum: [admin, coach] }
        teamId: { type: ["integer", "null"] }
    UpdateTeamInput:
      type: object
      properties:
        name: { type: string, minLength: 1 }
        shortName: { type: string, minLength: 1 }
        captainId: { type: integer }
    AttackInput:
      type: object
      required: [id, label, description, status, winnerTeamId]
      properties:
        id: { type: integer }
        label: { type: string }
        description: { type: string }
        status: { type: string }
        winnerTeamId: { type: ["integer", "null"] }
    CreateMatchInput:
      type: object
      required: [week, status, dateLabel, homeTeamId, awayTeamId, homeScore, awayScore, attacks]
      properties:
        week: { type: integer }
        status: { type: string }
        dateLabel: { type: string }
        homeTeamId: { type: integer }
        awayTeamId: { type: integer }
        homeScore: { type: integer }
        awayScore: { type: integer }
        attacks:
          type: array
          items: { $ref: "#/components/schemas/AttackInput" }
    UpdateMatchInput:
      type: object
      properties:
        week: { type: integer }
        status: { type: string }
        dateLabel: { type: string }
        homeTeamId: { type: integer }
        awayTeamId: { type: integer }
        homeScore: { type: integer }
        awayScore: { type: integer }
        attacks:
          type: array
          items: { $ref: "#/components/schemas/AttackInput" }
    Settings:
      type: object
      required: [seasonLabel, completedWeeks, totalWeeks]
      properties:
        seasonLabel: { type: string }
        completedWeeks: { type: integer }
        totalWeeks: { type: integer }
    UpdateSettingsInput:
      type: object
      properties:
        seasonLabel: { type: string, minLength: 1 }
        completedWeeks: { type: integer, minimum: 0 }
        totalWeeks: { type: integer, minimum: 1 }
    UpdatePlayerInput:
      type: object
      properties:
        name: { type: string, minLength: 1 }
        teamId: { type: integer }
        manOfMatch: { type: boolean }
    CreateQuestionInput:
      type: object
      required: [category, prompt, points, difficulty]
      properties:
        category: { type: string, minLength: 1 }
        prompt: { type: string, minLength: 1 }
        points: { type: integer, minimum: 1, maximum: 100 }
        difficulty: { type: string, minLength: 1 }
        isPublished: { type: boolean, default: true }
    CompetitionEventInput:
      type: object
      required: [playerId, category, points, reason]
      properties:
        playerId: { type: integer }
        category: { type: string }
        points: { type: integer, minimum: -10, maximum: 20 }
        reason: { type: string, minLength: 1 }