# lib/api-zod/src/generated/types/attackInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface AttackInput {
  id: number;
  label: string;
  description: string;
  status: string;
  /** @nullable */
  winnerTeamId: number | null;
}