# lib/api-zod/src/generated/types/createMatchInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AttackInput } from './attackInput';

export interface CreateMatchInput {
  week: number;
  status: string;
  dateLabel: string;
  homeTeamId: number;
  awayTeamId: number;
  homeScore: number;
  awayScore: number;
  attacks: AttackInput[];
}