# lib/api-zod/src/generated/types/createQuestionInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CreateQuestionInput {
  /** @minLength 1 */
  category: string;
  /** @minLength 1 */
  prompt: string;
  /**
     * @minimum 1
     * @maximum 100
     */
  points: number;
  /** @minLength 1 */
  difficulty: string;
  isPublished?: boolean;
}