# lib/api-zod/src/generated/types/loginInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface LoginInput {
  /** @minLength 1 */
  username: string;
  /** @minLength 1 */
  pin: string;
}