# lib/api-zod/src/generated/types/loginResult.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { LoginResultRole } from './loginResultRole';

export interface LoginResult {
  token: string;
  name: string;
  username: string;
  role: LoginResultRole;
  /** @nullable */
  teamId: number | null;
}