# lib/api-zod/src/generated/types/player.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { PlayerAttributes } from './playerAttributes';
import type { PlayerStats } from './playerStats';

export interface Player {
  id: number;
  name: string;
  teamId: number;
  teamName: string;
  role: string;
  avatarInitials: string;
  cardRating: number;
  cardToken: string;
  attributes: PlayerAttributes;
  stats: PlayerStats;
}